+++ src/App.tsx
@@ -4,11 +4,11 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import TopPage from "./pages/top-page/TopPage";
 import LandingPage from "./pages/landing-page/LandingPage";
 import RegularPage from "./pages/regular-page/RegularPage";
-import PostNewMeme from "./components/post-new-meme/PostNewMeme";
-import ThemeSwitcher from "./components/theme-switcher/ThemeSwitcher";
+import PostNewMeme from "./components/PostNewMeme/PostNewMeme";
+import ThemeSwitcher from "./components/ThemeSwitcher/ThemeSwitcher";
 import Error404 from "./pages/error404/Error404";
-import Footer from "./components/footer/Footer";
-import Navbar from "./components/navbar/Navbar";
+import Footer from "./components/Footer/Footer";
+import Navbar from "./components/Navbar/Navbar";
 
 function App() {
   useEffect(() => {
